fix(ci): bound mutation ownership to changed targets
