+++ backend/app/api/deps/authorization.py
@@ -8,7 +8,7 @@ import logging
 from typing import Annotated
 from uuid import UUID
 
-from fastapi import Depends, Header, Request, status
+from fastapi import Depends, Request, status
 from sqlalchemy.exc import SQLAlchemyError
 from sqlalchemy.ext.asyncio import AsyncSession
 
@@ -298,76 +298,6 @@ async def get_prepared_authorization_service(
     """Compose one request-local prepared service without taking commit ownership."""
     async with prepared_authorization_service(request, resolved, session) as service:
         yield service
-
-
-def require_submission_policy_mutation_key(
-    idempotency_key: Annotated[str | None, Header(alias="Idempotency-Key")] = None,
-) -> UUID:
-    """Validate manual policy replay custody before actor provisioning."""
-    if idempotency_key is None:
-        raise StructuredHTTPException(
-            status_code=422,
-            detail="Idempotency-Key must be a UUID",
-            error_code="validation_error",
-            error_message="Idempotency-Key must be a UUID",
-        )
-    try:
-        return UUID(idempotency_key)
-    except ValueError as exc:
-        raise StructuredHTTPException(
-            status_code=422,
-            detail="Idempotency-Key must be a UUID",
-            error_code="validation_error",
-            error_message="Idempotency-Key must be a UUID",
-        ) from exc
-
-
-async def require_submission_policy_human(
-    key: Annotated[UUID, Depends(require_submission_policy_mutation_key)],
-    result: Annotated[AuthVerificationResult, Depends(get_auth_verification_result)],
-) -> AuthVerificationResult:
-    """Conceal the public manual-policy surface from service principals."""
-    del key
-    if result.token.subject_kind != "human":
-        raise StructuredHTTPException(
-            status_code=404,
-            detail="Project authorization resource not found",
-            error_code="project_authorization_resource_not_found",
-            error_message="Project authorization resource not found",
-        )
-    return result
-
-
-async def submission_policy_authorization_actor(
-    request: Request,
-    result: Annotated[AuthVerificationResult, Depends(require_submission_policy_human)],
-    session: Annotated[AsyncSession, Depends(get_db_session)],
-    rate_control: Annotated[RateControlService, Depends(get_rate_control_service)],
-) -> ResolvedActor:
-    """Resolve a human only after the submission-policy key gate succeeds."""
-    return await resolve_authorization_actor(request, result, session, rate_control)
-
-
-async def get_submission_policy_prepared_authorization_service(
-    request: Request,
-    resolved: Annotated[ResolvedActor, Depends(submission_policy_authorization_actor)],
-    session: Annotated[AsyncSession, Depends(get_db_session)],
-):
-    """Compose submission-policy PREP from its dedicated admitted actor."""
-    async with prepared_authorization_service(request, resolved, session) as service:
-        yield service
-
-
-async def submission_policy_authorization(
-    key: Annotated[UUID, Depends(require_submission_policy_mutation_key)],
-    resolved: Annotated[ResolvedActor, Depends(submission_policy_authorization_actor)],
-    prepared: Annotated[
-        PreparedAuthorizationService,
-        Depends(get_submission_policy_prepared_authorization_service),
-    ],
-):
-    """Return the exact actor, key, and PREP service for manual policy mutation."""
-    return key, resolved, prepared
 
 
 @asynccontextmanager
